Delete old HostRepo and replace to database
